+++ tests/integration/test_otlp_logs.py
@@ -135,7 +135,8 @@ def test_otlp_logs_conversion(
                 "sentry.observed_timestamp_nanos": {
                     "stringValue": time_within(ts, expect_resolution="ns")
                 },
-                "sentry.payload_size_bytes": {"intValue": "358"},
+                "sentry.origin": {"stringValue": "auto.otlp.logs"},
+                "sentry.payload_size_bytes": {"intValue": "385"},
                 "sentry.severity_text": {"stringValue": "info"},
                 "sentry.span_id": {"stringValue": "eee19b7ec3c1b174"},
                 "sentry.timestamp_nanos": {
@@ -185,7 +186,7 @@ def test_otlp_logs_conversion(
             "org_id": 1,
             "outcome": 0,
             "project_id": 42,
-            "quantity": 358,
+            "quantity": 385,
         },
     ]
 
@@ -249,6 +250,7 @@ def test_otlp_logs_multiple_records(
                 "sentry.observed_timestamp_nanos": {
                     "stringValue": time_within(ts, expect_resolution="ns")
                 },
+                "sentry.origin": {"stringValue": "auto.otlp.logs"},
                 "sentry.payload_size_bytes": {"intValue": mock.ANY},
                 "sentry.severity_text": {"stringValue": "error"},
                 "sentry.span_id": {"stringValue": "eee19b7ec3c1b174"},
@@ -288,6 +290,7 @@ def test_otlp_logs_multiple_records(
                 "sentry.observed_timestamp_nanos": {
                     "stringValue": time_within(ts, expect_resolution="ns")
                 },
+                "sentry.origin": {"stringValue": "auto.otlp.logs"},
                 "sentry.payload_size_bytes": {"intValue": mock.ANY},
                 "sentry.severity_text": {"stringValue": "debug"},
                 "sentry.span_id": {"stringValue": "eee19b7ec3c1b175"},
@@ -337,6 +340,6 @@ def test_otlp_logs_multiple_records(
             "org_id": 1,
             "outcome": 0,
             "project_id": 42,
-            "quantity": 251,
+            "quantity": 305,
         },
     ]